pyproject.toml
[tool.poetry]
name = "mev-proxy"
version = "0.1.0"
description = "A simple proxy for the mev-eth protocol"
authors = ["ControlCplusControlV <44706811+ControlCplusControlV@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "mev_proxy"}]

[tool.poetry.dependencies]
python = "^3.10"
eth-ape = "^0.6.11"
colorama = "^0.4.6"
click = "^8.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
